requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
joblib==1.5.3
kafka-python==2.3.0
lightgbm==4.6.0
minio==7.2.20
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
py4j==0.10.9.9
pycparser==2.23
pycryptodome==3.23.0
pyspark==4.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.1
